Create a Notifier class template to match the Notifiable
